Tinh tong phan thuc - 1

+++ C++/CPP0103.cpp
@@ -6,7 +6,7 @@ using namespace std;
 int main() {
     int n;
     cin >> n;
-    float res = 0;
+    double res = 0;
     for (int i = 1; i <= n; i++) {
         res += 1.0 / i;
     }
